hw-3 is ready
